audio-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
scipy==1.13.1
librosa==0.10.2.post1
pyacoustid==1.3.0
pydub==0.25.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4

